Fix #15 - refactor css and introduce mixin
