+++ package.json
@@ -21,6 +21,8 @@
     "@types/react": "17.0.11",
     "autoprefixer": "^10.4.8",
     "eslint-config-next": "11.0.1",
+    "postcss-import": "^15.0.0",
+    "postcss-mixins": "^9.0.4",
     "postcss-nested": "^5.0.6",
     "prettier": "2.3.1",
     "typescript": "4.3.4"
