+++ src/components/footer.tsx
@@ -1,8 +1,8 @@
-import Linker, { LinkerProps } from './linker'
-import styles from '@src/styles/Linker.module.css'
+import styles from '@src/styles/Footer.module.css'
 import classNames from 'classnames/bind'
-
 const cx = classNames.bind(styles)
+
+import Linker, { LinkerProps } from './linker'
 
 interface FooterProps {
   tree: LinkerProps
@@ -10,7 +10,7 @@ interface FooterProps {
 
 export default function Footer({ tree }: FooterProps) {
   return (
-    <div className={cx('linker_container')}>
+    <div className={cx('root')}>
       <Linker {...tree} />
     </div>
   )
