+++ src/pages/_app.tsx
@@ -1,9 +1,10 @@
-import Head from 'next/head'
-import type { AppProps } from 'next/app'
 import '@src/styles/reset.css'
 import '@src/styles/variables.css'
 import '@src/styles/globals.css'
 import 'highlight.js/styles/monokai-sublime.css'
+
+import Head from 'next/head'
+import type { AppProps } from 'next/app'
 
 export default function MyApp({ Component, pageProps }: AppProps) {
   return (
