+++ src/styles/Linker.module.css
@@ -1,60 +1,51 @@
-.linker_container {
-  box-shadow: 0 2px 6px black;
-  margin-bottom: 20px;
-  padding: 10px 50px;
-  background-color: var(--javagray);
-  color: var(--header-font-color);
-  font-family: var(--content-font);
+.linker {
+  width: max-content;
+  display: flex;
+  flex-direction: column;
+  font-size: 18px;
 
-  .linker {
-    width: max-content;
+  .children {
+    margin-left: 1.9rem;
+  }
+
+  .linker_row {
     display: flex;
-    flex-direction: column;
-    font-size: 18px;
+    align-items: center;
+    margin: 3px 0;
 
-    .children {
-      margin-left: 1.9rem;
+    .arrow {
+      position: relative;
+      content: '';
+      display: inline-block;
+      width: 0.4em;
+      height: 0.4em;
+      border-right: 1px solid white;
+      border-top: 1px solid white;
+      transform: rotate(45deg);
+      margin-right: 0.5em;
+      transition: all 150ms cubic-bezier(0.25, 0.1, 0.25, 1);
+
+      &.isOpen {
+        transform: translateY(-3px) rotate(135deg);
+      }
     }
 
-    .linker_row {
-      display: flex;
-      align-items: center;
-      margin: 3px 0;
+    .linker_label {
+      padding: 8px 1rem;
+      border-radius: 10px;
 
-      .arrow {
-        position: relative;
-        content: '';
-        display: inline-block;
-        width: 0.4em;
-        height: 0.4em;
-        border-right: 1px solid white;
-        border-top: 1px solid white;
-        transform: rotate(45deg);
-        margin-right: 0.5em;
-        transition: all 150ms cubic-bezier(0.25, 0.1, 0.25, 1);
+      &.isFolder {
+        text-decoration: underline;
+        color: #b082c5;
 
-        &.isOpen {
-          transform: translateY(-3px) rotate(135deg);
+        &:before {
+          content: './';
         }
       }
 
-      .linker_label {
-        padding: 8px 1rem;
-        border-radius: 10px;
-
-        &.isFolder {
-          text-decoration: underline;
-          color: #b082c5;
-
-          &:before {
-            content: './';
-          }
-        }
-
-        &.isSelected {
-          background-color: var(--header-font-color);
-          color: var(--javagray);
-        }
+      &.isSelected {
+        background-color: var(--header-font-color);
+        color: var(--javagray);
       }
     }
   }
