+++ src/styles/common.pcss
@@ -1,0 +1,28 @@
+@define-mixin root {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+
+  .top_container {
+    width: 100%;
+    max-width: 920px;
+    min-height: 100vh;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+
+    @mixin-content;
+  }
+}
+
+@define-mixin card {
+  box-shadow: 0 2px 6px black;
+  margin-bottom: 20px;
+  padding: 32px 53px 17px 53px;
+
+  color: var(--content-font-color);
+  background-color: var(--content-background);
+
+  @mixin-content;
+}
